@@ -12,9 +12,9 @@ public final class ApachePoiImportUtils {
 	 *         This is all about formulas : if the given cell is of type {@link Cell#CELL_TYPE_FORMULA}, this returns the return type of the formula.
 	 */
 	public static CellType getCellActualValueType(Cell cell) {
-		CellType cellType = cell.getCellTypeEnum();
+		CellType cellType = cell.getCellType();
 		if (cellType.equals(CellType.FORMULA)) {
-			return cell.getCachedFormulaResultTypeEnum();
+			return cell.getCachedFormulaResultType();
 		} else {
 			return cellType;
 		}
