@@ -2,12 +2,6 @@ package org.iglooproject.jpa.more.config.spring;
 
 import java.util.Collection;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectMapper.DefaultTyping;
-
 import org.iglooproject.jpa.more.business.CoreJpaMoreBusinessPackage;
 import org.iglooproject.jpa.more.business.task.dao.IQueuedTaskHolderDao;
 import org.iglooproject.jpa.more.business.task.dao.QueuedTaskHolderDaoImpl;
@@ -16,15 +10,27 @@ import org.iglooproject.jpa.more.business.task.service.IQueuedTaskHolderManager;
 import org.iglooproject.jpa.more.business.task.service.IQueuedTaskHolderService;
 import org.iglooproject.jpa.more.business.task.service.QueuedTaskHolderManagerImpl;
 import org.iglooproject.jpa.more.business.task.service.QueuedTaskHolderServiceImpl;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectMapper.DefaultTyping;
+import com.fasterxml.jackson.databind.jsontype.impl.LaissezFaireSubTypeValidator;
 
 @ComponentScan(basePackageClasses = { CoreJpaMoreBusinessPackage.class })
 public abstract class AbstractTaskManagementConfig {
 	
 	public static final String OBJECT_MAPPER_BEAN_NAME = "queuedTaskHolderObjectMapper";
 
+	/**
+	 * Base configuration for task serialization. We use a permissive deserializer config as data are internally
+	 * managed by application.
+	 */
 	@Bean(name = OBJECT_MAPPER_BEAN_NAME)
 	public ObjectMapper queuedTaskHolderObjectMapper() {
-		return new ObjectMapper().enableDefaultTyping(DefaultTyping.NON_FINAL);
+		return new ObjectMapper().activateDefaultTyping(
+				LaissezFaireSubTypeValidator.instance,
+				DefaultTyping.NON_FINAL);
 	}
 
 	@Bean
