@@ -1,8 +1,8 @@
 package org.iglooproject.test.jpa.batch;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
 
 import java.util.Collections;
 import java.util.List;
