@@ -1,11 +1,15 @@
 package org.iglooproject.jpa.batch.executor;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.Date;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
+import org.iglooproject.jpa.batch.processor.ThreadedProcessor;
+import org.iglooproject.jpa.batch.runnable.IBatchRunnable;
+import org.iglooproject.jpa.batch.util.ProcessorProgressLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
@@ -17,10 +21,6 @@ import org.springframework.transaction.support.TransactionCallback;
 import org.springframework.transaction.support.TransactionTemplate;
 
 import com.google.common.collect.Lists;
-
-import org.iglooproject.jpa.batch.processor.ThreadedProcessor;
-import org.iglooproject.jpa.batch.runnable.IBatchRunnable;
-import org.iglooproject.jpa.batch.util.ProcessorProgressLogger;
 
 @Component
 @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
@@ -145,9 +145,9 @@ public class MultithreadedBatchExecutor extends AbstractBatchExecutor<Multithrea
 		if (duration < 1000) {
 			sb.append(String.format("in %1$d ms", duration));
 		} else if (duration < 60000){
-			sb.append(String.format("in %1$s s", BigDecimal.valueOf(duration / 1000f).setScale(3, BigDecimal.ROUND_HALF_UP).toString()));
+			sb.append(String.format("in %1$s s", BigDecimal.valueOf(duration / 1000f).setScale(3, RoundingMode.HALF_UP).toString()));
 		} else {
-			sb.append(String.format("in %1$s mn", BigDecimal.valueOf(duration / 60000f).setScale(2, BigDecimal.ROUND_HALF_UP).toString()));
+			sb.append(String.format("in %1$s mn", BigDecimal.valueOf(duration / 60000f).setScale(2, RoundingMode.HALF_UP).toString()));
 		}
 		if (e != null) {
 			// Log this as info anyway, since error handling is done elsewhere
