@@ -1,8 +1,8 @@
 package test.wicket.more.model;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
 
 import java.util.Arrays;
 import java.util.Iterator;
